Remove Random Stencil Clear On Left Leg (#1955)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

I don't know why this was being done, but it didn't seem to do anything
and I know for a fact it was causing issues with shaders that were
applied on all humanoid layers, so... "Imaginary Technique: Shitcode
Removal" go.

If anything breaks, it's this.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [X] Check that the game didn't incinerate itself.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/ff7f41e5-827e-4fdf-98e2-b5885aadf467)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

no cl

Signed-off-by: Eris <eris@erisws.com>

(cherry picked from commit e674f782e24da7a7825520ca650473f155e9e4c1)