Saltern Fixes (#1937)
# Description

My mappers somehow all at once decided that removing all shuttles was a
good idea. This caused problems because the mining shuttle sometimes
doesn't dock to the station correctly, and when it happens a shuttle is
needed to go retrieve it. Making it so that cargo doesn't have either
cargo shuttle or pathfinder makes this impossible. @Ichaie
@Mike32oz @OldDanceJacket make sure that in future map updates, that
your maps have at least one pilotable shuttle, either the cargo shuttle
or the Pathfinder. Ideally, you should be using the new Shuttle Spawning
Airlocks, such as AirlockExternalGlassShuttlePathfinderFilled

These airlocks spawn a ship already docked to the station, similar to
how the escape pods work.

Also make sure that the doors to Supermatter cores are using High
Security Engineering Locked doors, these airlocks are special in that
they don't allow smallmobs like mice to crawl underneath, and are harder
to hack. That first point is extremely vital because shitters will take
mouse ghostroles and sprint into the supermatter to try and detonate it.

Anyways this PR adds a dock for the Pathfinder to Saltern, as well as
fixes some issues with Saltern's Supermatter.

# Changelog

:cl:
- add: Saltern now has a dock for the Pathfinder, which will now start
the round already docked to the station if nothing goes wrong.
- tweak: Made Saltern's supermatter much less "Explodey"

(cherry picked from commit d9fd6012285570a0d860ce86df480d85c0671ba6)