More Loadout Items (#1967)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

This PR adds the following items to the loadout menu.

*  Expensive Coat
* Both tacticool outfits (jumpskirt and suit)
* Full tacticool maid set
* Witch Hat and robes
* Luxury Pen
* A few plushies

This allows easier access to less used items and provides more
customization to characters.
<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/df3fc234-12f2-4f66-96ce-b8354bd4547e)

![image1](https://github.com/user-attachments/assets/1e3c9774-8bc9-4a17-bb6f-3a37e0bf69ae)

![image4](https://github.com/user-attachments/assets/f758e7ff-f352-4bf7-a098-2455f265d452)

![image5](https://github.com/user-attachments/assets/33b9a66d-3f38-4595-840b-7afbd7b726b0)

![image6](https://github.com/user-attachments/assets/21dae742-edb2-4863-bb37-8a43559b5334)

![image2](https://github.com/user-attachments/assets/12b5d46c-807a-467b-a264-cdb4830f64e4)

![image3](https://github.com/user-attachments/assets/334ccb30-36b4-4f00-8f40-fa3d598e7b06)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Tonk
- add: Added tacticool outfits, expensive coat, witch hat and robe to
the loadout menu.
- add: Added  plushies and luxury pen to loadout items.

---------

Signed-off-by: Tonk-GCR <190437025+Tonk-GCR@users.noreply.github.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>
(cherry picked from commit 0262883792f15c62d6d0130613a7a90d65ba3a05)