Shadowkin Blood Change (#1904)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Replaces Shadowkin's blood type with black blood, to fit their
pre-existing organ set
<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Add reagent Black Blood
- [x] Change shadowkin blood type and injury color
- [x] Lock in descriptions and exact hex code
- [x] Add artifact trigger

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

![420641667-bde592ce-d26e-4656-ab24-14d85fee4741](https://github.com/user-attachments/assets/1454e659-3ae6-4e07-98fa-f2bc88fd238b)

![420641642-f51cafd2-b1ee-4e64-8e5e-80923feb7e09](https://github.com/user-attachments/assets/1600acf7-7d41-4c54-914b-775fbd29d20e)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:

- tweak: Changed shadowkin blood type to black blood, to better fit
their organs' color palette

(cherry picked from commit 33de108fb39d09e2eff5cce586755671007a6746)