Port Changelings From Goobstation (Funky PR 387 Included) (#1855)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Finally gets around to porting over Changelings from Goobstation, as
well as a *certain evil PR* from FunkyStation (with the fixes it comes
with).
---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Wait for #1856 to be merged
- [x] Wait for #1860 to be merged
- [x] Fix broken code to make it actually compile (right now is just
porting prototypes, code, and locale)
- [X] Port That Funky PR I Mentioned Earlier
- [] Throw bricks at the codebase until it stops failing tests
- [X] Maybe do some local testing

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/04473aef-d076-4dd0-a700-0c656428b88a)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Changelings have been ported!

---------

Signed-off-by: Eris <erisfiregamer1@gmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
(cherry picked from commit ee4f7aa7097f260d07af3ab0583a2a5c14a38c74)