Dreadnought "Power Armor" / Hardsuit for the Boring People. (#1950)
This simply ports over the Dreadnought suit from Goob LRP with a change
or two to it, I recommend looking at the Goob LRP PR for more
information.

https://github.com/Goob-Station/Goob-Station/pull/1973

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

The funny dreadnought suit that fucking kills you slowly.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Task
- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://i.imgur.com/FyQXheo.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Cybersun has gotten desperate after hearing several of their
juggernaut suit users got robusted, so they made whatever the fuck this
is, the fucking hardsuit that kills you slowly.
- add: Funny code that allows for damage over time to happen, isn't that
so cool.

---------

Signed-off-by: BloodfiendishOperator <141253729+Diggy0@users.noreply.github.com>
Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>
(cherry picked from commit c8256ab6a61208325fa502b427e3a0bda340cca6)