Book Press (#1977)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

A machine for the librarian that an print most of the existing books in
the game.
There are two goals to this:
- Give the Librarian something to do
- Add an incentive for people to make more books, because now they'll be
reliably obtainable.

Books added to the Book Press can be fiction, but also guidebooks on
more advanced topics not contained in the in-game guide book, or recipe
books, guidelines, or any other useful information we want to have
available in-character.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Machine Sprite
- [x] Add Recipes

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added Book Press. Librarians rejoice!

(cherry picked from commit 96514d368060de13776e7baf6ac07b554d96df52)