Contractors V1 (#2030)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Contractors v1 is the first version of a system that will allow you to
set a nationality and an employer to your character. Initially, this
will determine the availability of some loadout items, jobs and traits,
but this is built to expand upon later.

As of this moment, the PR will let you select a nationality, an employer
and a lifepath. Nationalities give you a language and a passport, while
the other two don't do a lot yet. (except Command can only be NanoTrasen
and most other jobs can't be unemployed)

The passport functions, and the CharacterRequirements do as well.

there's still a lot more that can be done--tying jobs to certain
employers, items or traits to nationalities or lifepaths, but the reason
I want to merge it before that's done is primarily so that I don't need
to be the sole person working on it anymore. The C# is done, the rest is
YAML, and we have a bunch of competent YAML warriors who will do great
stuff with this, no doubt.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Create Nationality and Employer Prototype code
- [x] Create Nationality and Employer Character Requirements
- [x] Add both fields to the Character Creation menu
- [x] Create Nationality and Employer Prototypes
- [x] Create Nationality and Employer Prototypes YAML
- [x] Add requirements to a few jobs
- [ ] Add requirements to a few traits
- [ ] Add requirements to a few loadout items
- [x] create a passport item that can be opened and closed
- [x] Give each nationality a passport item valid for their species

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/9c44cf11-766b-4368-af69-8655048e992a)

https://github.com/user-attachments/assets/9e61aed8-2e07-4d44-89c7-595a170df8c7

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Contractors. Be sure to edit your character to set an employer
and nationality!
- add: Nationality, Lifepath, and Employer have been added to character
creation. These don't currently do much except spawn a passport
containing information about where your character is from and who they
are, but they are fully integrated with Character Requirements, and by
extension can be used for and alongside Traits, Loadouts, Jobs,
Antagonists, etc.

---------

Signed-off-by: Timfa <timfalken@hotmail.com>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>

(cherry picked from commit c3b12c62ee3bd226b57022690d1b24cb7cf54e3f)