Multiclass Trait Loadout Changes (#2143)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Makes all the weapons cost more points
Increase Cost of Insuls and Reduced the Cost of Holofan to 1
I will do another pr to add clothes and the like later

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] stuff
---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Sec Weapons for The Restricted Gear Trait now Cost a Base Line
of 8 points
- tweak: Weapons in General is More Expensive
- tweak: Insuls are Slightly More Expensive
- tweak: Holofan cost 1 now
- Removed: E-G4 from Restricted Gear

---------

Signed-off-by: WazuraWasTaken <dumplingyen@gmail.com>
(cherry picked from commit 24364951d2b6c6c1ba35bef4290cbce81e9a5806)