Port Thaven From DeltaV, Who Ported It From Impstation (#2132)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

https://github.com/DeltaV-Station/Delta-v/pull/2948 and a bunch of other
PR's

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] fix markings causing a crash
- [x] fix displacement maps causing a crash

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/ad449fc4-93c7-47d4-8db7-c53da765aa51)

https://github.com/user-attachments/assets/033bec46-24c5-44ad-8e5c-7aae2ed85b03

https://github.com/user-attachments/assets/6fd647b5-2ee0-45e6-a124-9b90c35e2153

![image-71](https://github.com/user-attachments/assets/bdb7b129-a1b4-445d-9b1a-fa884b429ad4)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Port Thaven from DeltaV, who ported it from Impstation

(cherry picked from commit 86a8e4d940eac038505c66a1649b376f6956cade)