Senior Roles Stuff (Ft. CMO and Other Loosely Related Changes)  (#2155)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This PR adds better tools to take for certain senior roles, seperate the
medkits and survival boxes in the loadout options so they can be taken
seperately, make and give the CMO a worse version of the multitool,
which does everything at 1.25x speed (for comparison the normal
multitool does everything at 2x speed) and also let the CMO take their
armor rig in their loadout. (with it being pre-filled)
Also went ahead and added descriptions to the medical multitool. (As it
would be silly to not do so if I have given one to the one given to CMO)

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Seperate medkits and survival boxes in loadouts
- [x] Make the field surgery kit: a combat medkit that is instead filled
with advanced surgical tools, bone gel, and some other things
- [x] Give Senior Physician the option to choose between a bluespace
beaker and the field surgery kit
- [x] Give Senior Engineer the experimental welding tool to take in
their loadout
- [x] Give CMO their armor rig to take in their loadout
- [x] Fill the armor rig that CMO takes
- [x] Make a worse version of the medical multitool that the CMO can
take in their loadout
- [x] Add descriptions for both the CMO multitool and the normal one

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Base Profile Screenshot 2025 04 06 - 21 02 13
99](https://github.com/user-attachments/assets/abb39520-9ce0-4d84-aba8-6b0688fad02a)
![Base Profile Screenshot 2025 04 06 - 21 02 28
29](https://github.com/user-attachments/assets/60105083-9b9b-4162-829d-18036d7b81a3)

https://github.com/user-attachments/assets/aad9316c-f8ad-4be9-9589-4fdd9f3e4586

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added a worse version of the medical multitool that the CMO
starts with (taken in loadouts)
- add: Description for the (normal) medical multitool (as I was adding
one for the proto one CMO gets so might as well)
- add: Added the a filled version of the CMO's armor rig, available in
their loadout (The goodies inside are: 2x of each adv topical, 2x blood
bags, 1x omnizine bootle, 1x medipen box, which is a blue cig box with
2x rad, poison and space pens, as well as a brute, burn and emergency
pen)
- add: Added the: Experimental welder, for senior engineer (egnineer),
and the bluespace beaker OR field surgical kit (combat medkit but with
adv surgical tools + bone gel, chloral hydrate pills and an emergency
medipen) in their respective loadouts
- tweak: Seperated the loadout medkits and survival boxes so that you
can take both

---------

Co-authored-by: VMSolidus <evilexecutive@gmail.com>
(cherry picked from commit 01e0d009dfb01d936c5765211498f4a581e0653c)