Plasma Window Buffs (#2195)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Large buff to plasma windows so that epi doesn't instantly cease to
exist the moment an artifact blows up.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Plasma windows now take 50% less explosion damage
- tweak: Reinforced plasma windows now take 80% less explosion damage

---------

Signed-off-by: EctoplasmIsGood <109397347+EctoplasmIsGood@users.noreply.github.com>
Co-authored-by: Your Name <EctoplasmIsGood@users.noreply.github.com>

(cherry picked from commit 898ae13dfe3686892205671196919ae63f4cdae5)