The Nuke Cleanade (#2194)
Because I hate tiders

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# The great nuke cleanade.

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Works way better than a cleanade, the recipe being 100 bleach, one soap,
and one cleanade into the microwave for 10 seconds.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Diggy
- add: Cleanadenuke grenade has been added, made with 100 units of
bleach, one soap and a cleanade, put it in the microwave for 10 seconds
and there you have it, enjoy flooding a room with space cleaner foam.

---------

Signed-off-by: BloodfiendishOperator <141253729+Diggy0@users.noreply.github.com>
Co-authored-by: Timfa <timfalken@hotmail.com>
(cherry picked from commit d5dcfc89621c5cb0f8dbe9380632d3cf939bbb2f)