Maint Drone Language (#2208)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Allowes drones to speak to one another with "Drone" Moreover they now
understand but cannot speak TauCeti!

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] make drone lang
- [x] allow drones to speak dronelang
- [x] make drones beep when speaking

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/362fc2a0-3129-4b29-a665-190bb9ff74fa)

![image](https://github.com/user-attachments/assets/d73aa911-8616-4120-afdc-f5e7b437239c)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Maint drones can now speak to other maint drones via the Drone
language!

---------

Signed-off-by: RocketBoss360 <rocketboss360@gmail.com>
(cherry picked from commit 69b9710e7757b3fe2fff76e890915a35806f97f8)