Initial Corporate Requirements (#2213)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

NanoTrasen is a megacorporation primarily focused on advanced starship
engine development and cutting-edge phoron research. Due to the highly
specialised nature of its core operations, it is both more
cost-effective and operationally efficient to outsource many of the
station's non-core roles—ranging from food service and custodial duties
to certain logistical and clerical tasks—to contracted personnel and
external agencies.

I've made this PR to implement the requirements reflecting this
organisational structure, ensuring the game world reflects NanoTrasen’s
corporate focus and reliance on outsourced labour where appropriate.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/82a2ec28-c27c-4551-abe0-3bbc5eb0ba4e)

![image](https://github.com/user-attachments/assets/fdc9eb83-7448-4e43-a23d-9f42c5118cee)

![image](https://github.com/user-attachments/assets/4df7eca9-ed4f-482a-b120-c6a6ebe4c509)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added employer requirements to most roles. CHECK YOUR CHARACTER'S
JOB PREFERENCES AND EMPLOYER!

(cherry picked from commit e3e83ba7fea300839489c208dbda239e68096bbd)