Woe, Cybernetic Appendage Trait Balancing Be upon Ye. (#2255)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This PR tries to better balance bionic arms and comments out speed legs
FOR NOW.

For Bionic arms, it restricts them to only be accessible to security,
command, dignitaries and paramedics, while also knocking down the cost
back to 8, as it used to be. (this is because it is now only available
to a select few and not everyone)
On top of that, I added a new trait, Prybar Prosthetics, which are
cybernetic arms that can pry open UNPOWERED doors (so it's a crowbar).
This one is much cheaper than bionic arm (3 points) and is available to
everyone besides prisoners. (Yes it replaces your limbs with a new
cybernetic arm with this function, the DX-1 Pryarms)

For speed legs, like previously stated, they're getting commented out.
This is only temporary, until someone can figure out a way to rebalance
them (speed legs as a whole I mean, not the trait itself.) so they
aren't just a permament increase to speed. Speed is too good, look at
what harpies pay for a 10% increase while sprinting.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Base Profile Screenshot 2025 04 16 - 17 49 18
97](https://github.com/user-attachments/assets/ac8d502c-249b-401a-baf0-01904c7da1d5)
![Base Profile Screenshot 2025 04 16 - 17 49 30
79](https://github.com/user-attachments/assets/8ef903e5-1377-48a9-8847-90f16cf07ef5)
![Prybar Prosthetics
working](https://github.com/user-attachments/assets/d397da9e-33e5-452d-aefd-a3eb90f55a98)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added the Pryarm Prosthetics trait, that replaces your arms with
the DX-1 Pryarms, which let you pry open unpowered doors just as you
would with a crowbar.
- tweak: Made bionic arm cost 8 points again and be restricted to
command, security, dignitaries and paramedics only.
- remove: Speed legs trait go bye bye for now.

(cherry picked from commit 980ba8139fd78eaeebab464bac985cd34fe0ead0)