Medipen Case + Old Bruiz Recipe Restoration (#2256)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Smol PR that replaces the medipen box in CMO's (filled) webbing with the
medipen case. (Same thing but now uses the syringe case as the parent
entity, and has a new name n description)
Also reverts the bruizine recipe back to the old one because it's
supposed to be harder to make :godo:

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Base Profile Screenshot 2025 04 16 - 19 49 52
67](https://github.com/user-attachments/assets/c524e061-6976-4622-9be3-81f187846ce2)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Changed the medipen box used in the CMO's filled webbing into
the medipen case.
- tweak: Reverted bruizine recipe back to the old one.

(cherry picked from commit dbc23a263aad98d11323bd6f857e4106b139c920)