Pebble 3.0 (#2261)
# Pebble 3.0

This is basically a complete remap of Pebble, which is an old old old
Nyano map that has shifted hands multiple times and had a number of
individuals work on it. This station has had problems for years, and
beyond just it's age just a set of fundamental issues that had never
been addressed.
With the assistance of Nyawul, Plyush, Zeta, and a number of others --
This low pop map should now no longer be a contender for one of the
worst stations in the game.

---

# Changelog:

_-Every firelock, air alarm, and camera has been redone._
_-Medical and Security have been completely redone._
_-Cargo, Service, Bridge, and Epi have seen major overhauls bordering on
being completely redone._
_-Engineering has been adjusted._
_-Maints has been nearly tripled._
_-The station has been rebalanced to be a far better experience for an
antag(windows have been cut down, allowing for blind spots). I cannot
stress how bad it was for antags previously, cameras were in the
freezer._
_-Magistrate, Boxer, and Administrative Assistant have been removed from
the map._

---

:cl: ODJ
- tweak: Pebble updated to version 3.0, please enjoy.

---------

Signed-off-by: OldDanceJacket <98985560+OldDanceJacket@users.noreply.github.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>

(cherry picked from commit e834014d7452203fdddc89ad6c7589f156f3eb96)