Crew Monitor Map Fixes (#2302)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Description.

---
This doesn't address the root issue which is that this doesn't happen
automatically, but this will make it so that players don't have to fix
the crew monitor themselves round start.

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/71fa4cb7-b6d1-487f-8efd-8f31eea8d2a1)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- fix: The Crew Monitor server now works on Anchor, Asterisk, Core,
Hive, Lambda, and Pebble roundstart.

Co-authored-by: VMSolidus <evilexecutive@gmail.com>
(cherry picked from commit 89ddb9305ce57249526e82ec0350a0b537f7a401)