The Magmantle! (#2295)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This PR adds the Magmantle, a mantle that functions like magboots, but
slightly worse (like 15% more slowdown, but it does provide 20%
temperature protection like winter boots or scarves). Its intended for
species without the feet slot, which are harpy, diona and arachne. It is
available in the species tab of loadouts, for 6 points. Only Security,
Command, Dignitaries, Engineering, Salvies and Parameds can take it, if
they are a harpy, diona or arachne.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![mmantle](https://github.com/user-attachments/assets/782bef02-cba5-4d69-86f0-eece4aa3ea1a)

![mmantledesc](https://github.com/user-attachments/assets/455b3ede-8467-4aa9-a668-b266a2abde38)
![Magmantle working, speed modifiers
showcase](https://github.com/user-attachments/assets/bf03525d-3d8a-410b-bd3c-e1a9cdb0d1d5)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added the Magmantle, a mantle that functions as worse magboots on
the neck slot (15% more slow, but they give 20% temperature protection
like winter boots or scarves). It is available in the species section of
loadouts, for all the species that do not have a feet slot. (Harpy,
Diona, Arachne)

---------

Co-authored-by: VMSolidus <evilexecutive@gmail.com>
(cherry picked from commit 2e0bc38c63ac58ec70735287bd0a6954cea99562)